requirements.txt
Flask==2.3.3
Flask-RESTx==1.3.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
celery==5.3.4
redis==5.0.1
SQLAlchemy==2.0.48
Werkzeug==2.3.7
python-dotenv==1.0.0
pyotp==2.9.0
zxcvbn>=4.4.25
qrcode==7.4.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
dnspython==2.4.2
whois==0.9.25
WeasyPrint==60.2
selenium==4.15.2
webdriver-manager==4.0.1
Pillow==10.1.0
langdetect==1.0.9
gunicorn==21.2.0
python-socketio>=5.0.2
python-engineio>=4.11.0
python-dateutil==2.8.2
marshmallow==3.20.1
APScheduler==3.10.4
numpy>=1.24.0